changed structure.
