--- evals/evaluation/crud_rag/requirements.txt
@@ -1,2 +1,0 @@
-evaluate
-rouge_score
