+++ evals/evaluation/rag_eval/evaluator.py
@@ -8,7 +8,8 @@ import os
 import requests
 from tqdm import tqdm
 
-from .metrics import LLM_score, bleu_score, rougeL_score
+from .metrics import LLM_score
+from evals.metrics import bleu_score, rougeL_score
 
 
 class Evaluator:
