+++ evals/metrics/__init__.py
@@ -2,3 +2,5 @@
 # -*- coding: utf-8 -*-
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
+
+from .utils import bleu_score, rougeL_score
